Perf tweaks

+++ tests/test_import_sesar_records.py
@@ -14,15 +14,18 @@ def test_local_records():
     for file in os.listdir(dir):
         if file.startswith(ISAMPLES_FILE_PREFIX):
             continue
+        filename_no_ext = os.path.splitext(file)[0]
+        dest_path = dir + ISAMPLES_FILE_PREFIX + filename_no_ext + ".json"
+        if os.path.exists(dest_path):
+            continue
         full_path = os.path.join(dir, file)
         with open(full_path) as source_file:
             source_record = json.load(source_file)
             try:
                 transformer = SESARTransformer(source_record)
                 transformed_to_isamples_record = transformer.transform()
-                filename_no_ext = os.path.splitext(file)[0]
                 with open(
-                    dir + ISAMPLES_FILE_PREFIX + filename_no_ext + ".json", "w"
+                    dest_path, "w"
                 ) as outfile:
                     json.dump(transformed_to_isamples_record, outfile)
             except:
